requirements.txt
GitPython>=2.1.5
requests>=2.22.0
